requirements.txt
argon2==0.1.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cryptography==37.0.1
defusedxml==0.7.1
distro==1.7.0
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-configurations==2.3.2
django-crispy-forms==1.14.0
django-registration==3.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
oauthlib==3.2.0
paramiko==2.10.4
pkg_resources==0.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python3-openid==3.2.0
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.9
websocket-client==0.59.0
